Add options.packages.find section to whitelist the namespace package
Without a white list the test package was being added to the wheel.

Correct the exported README.md so the results from running the
code in REDME.org are included.

Add How to build and install the project.

Add section on what can go wrong by using the find directive in place
of the find_namespace directive